+++ packages/devops/lang/en/health.php
@@ -5,4 +5,5 @@ return [
         'label' => 'No failed background jobs',
         'message' => '{0} No failed background jobs|{1} 1 failed background job|[2,*] :count failed background jobs',
     ],
+    'check_failed_mail_subject' => 'Health check failed on :url',
 ];